voice - suspend keyword recognition when focus lost (#205336)
* speech - bind stop keyword detection to window focus

* .

* .

* .
