+++ src/vs/platform/auxiliaryWindow/electron-main/auxiliaryWindows.ts
@@ -5,7 +5,6 @@
 
 import { BrowserWindowConstructorOptions, HandlerDetails, WebContents } from 'electron';
 import { Event } from 'vs/base/common/event';
-import { Schemas, VSCODE_AUTHORITY } from 'vs/base/common/network';
 import { IAuxiliaryWindow } from 'vs/platform/auxiliaryWindow/electron-main/auxiliaryWindow';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 
@@ -29,8 +28,4 @@ export interface IAuxiliaryWindowsMainService {
 	getLastActiveWindow(): IAuxiliaryWindow | undefined;
 
 	getWindows(): readonly IAuxiliaryWindow[];
-}
-
-export function isAuxiliaryWindow(webContents: WebContents): boolean {
-	return webContents?.opener?.url.startsWith(`${Schemas.vscodeFileResource}://${VSCODE_AUTHORITY}/`);
 }
