+++ src/vs/workbench/contrib/speech/browser/speech.contribution.ts
@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { InstantiationType, registerSingleton } from 'vs/platform/instantiation/common/extensions';
-import { ISpeechService, SpeechService } from 'vs/workbench/contrib/speech/common/speechService';
+import { ISpeechService } from 'vs/workbench/contrib/speech/common/speechService';
+import { SpeechService } from 'vs/workbench/contrib/speech/browser/speechService';
 
 registerSingleton(ISpeechService, SpeechService, InstantiationType.Delayed);
